move api auth classes to their own package
