+++ corehq/apps/api/accounting.py
@@ -4,7 +4,8 @@ from corehq.apps.accounting.models import Feature, FeatureRate, SoftwarePlanVers
     BillingAccount, BillingContactInfo, Currency, PaymentRecord, SoftwareProductRate, \
     SoftwareProduct, SoftwarePlan, DefaultProductPlan, CreditAdjustment, Subscription, CreditLine, Subscriber, \
     SubscriptionAdjustment, BillingRecord, Invoice
-from corehq.apps.api.resources.v0_1 import CustomResourceMeta, AdminAuthentication
+from corehq.apps.api.resources.auth import AdminAuthentication
+from corehq.apps.api.resources.v0_1 import CustomResourceMeta
 from tastypie import fields
 from django_prbac.models import Role
