+++ corehq/apps/api/resources/v0_5.py
@@ -15,11 +15,9 @@ from django.core.urlresolvers import reverse
 
 from tastypie import fields
 from tastypie.bundle import Bundle
-from corehq.apps.api.resources.v0_1 import (
-    RequirePermissionAuthentication,
-    AdminAuthentication,
-    CustomResourceMeta,
-)
+
+from corehq.apps.api.resources.auth import RequirePermissionAuthentication, AdminAuthentication
+from corehq.apps.api.resources.v0_1 import CustomResourceMeta
 from corehq.apps.api.util import get_obj
 from corehq.apps.app_manager.models import Application
 from corehq.apps.domain.models import Domain
