+++ corehq/apps/locations/resources/v0_5.py
@@ -4,7 +4,7 @@ from tastypie.constants import ALL
 from tastypie.resources import ModelResource
 
 from corehq.apps.api.resources import HqBaseResource
-from corehq.apps.api.resources.v0_1 import DomainAdminAuthentication
+from corehq.apps.api.resources.auth import DomainAdminAuthentication
 from corehq.util.view_utils import absolute_reverse
 
 from ..models import SQLLocation, LocationType
