+++ custom/ewsghana/resources/v0_1.py
@@ -1,7 +1,7 @@
 import json
 
+from corehq.apps.api.resources.auth import LoginAndDomainAuthentication
 from corehq.apps.api.util import get_object_or_not_exist
-from corehq.apps.api.resources.v0_1 import LoginAndDomainAuthentication
 from corehq.apps.locations.models import SQLLocation, Location
 from corehq.apps.locations.resources.v0_1 import LocationResource
 from corehq.util.quickcache import quickcache
